# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Travellers app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of travellers app"
  },
  "servers": [
    {
      "url": "https://travellers-backend.onrender.com"
    },
    {
      "url": "http://localhost:3001"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/users/update-info": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Update user profile information",
        "operationId": "updateUserInfo",
        "description": "Update the current user profile information with a description",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "Привіт всім. Люблю подорожувати та ділитись своїми історіями"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User information successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated"
                    },
                    "data": {
                      "type": "object",
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/user"
                        },
                        {
                          "properties": {
                            "savedStories": {
                              "type": "array",
                              "example": [
                                "68498236a100312bea079029",
                                "694a8c750cbe696229c7bb24"
                              ]
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time",
                              "example": "2025-12-23T12:19:07.587Z"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time",
                              "example": "2025-12-26T05:13:44.469Z"
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string",
                  "example": "Validation error"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "example": {
              "status": 400,
              "message": "BadRequestError",
              "data": {
                "message": "Error text"
              }
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "message": {
                  "type": "string",
                  "description": "Error definition"
                },
                "data": {
                  "type": "object",
                  "description": "list of possible errors",
                  "properties": {
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "example": {
              "status": 401,
              "message": "UnauthorizedError",
              "data": {
                "message": [
                  "Session not found",
                  "Access token expired"
                ]
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "description": "list of possible errors",
                  "properties": {
                    "message": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "example": {
              "status": 404,
              "message": "NotFoundError",
              "data": {
                "message": [
                  "Route not found",
                  "User not found",
                  "Story not found"
                ]
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "694a88bbcbc8579663486147"
          },
          "name": {
            "type": "string",
            "example": "User Name"
          },
          "description": {
            "type": "string",
            "example": "Hello everyone, I am happy to share my stories."
          },
          "avatar": {
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "publicStories": {
            "type": "array",
            "example": [
              "68498236a100312bea079029",
              "694a8c750cbe696229c7bb24"
            ]
          }
        }
      },
...
    }
  }
}